requirements.txt
requests
ffmpeg-python
pydub
fuzzywuzzy
instagrapi
tiktok_uploader
opencv-python
